refactor!: rollback the change with TerminatedPlan
Using distinct store key prefix for active plans and terminated
plans was to prevent farming plan spamming, but that is now
mitigated by limiting the number of private plans.
Now, having distinct prefix for terminated/non-terminated plans
causes query endpoints to be separated too, and also requires more work
on CLI side.
It'd be better to rollback the change.